feat: add LegacyCalendar component without native dependencies
This commit introduces a new LegacyCalendar component that allows users to use the calendar without installing native dependencies (react-native-gesture-handler, react-native-infinite-pager, react-native-reanimated).

## Changes

### New Components
- **CalendarContainerCore**: Core calendar logic without native dependencies
- **LegacyCalendar**: Wrapper component for legacy usage via `react-native-big-calendar/legacy`

### Refactored Components
- **CalendarContainer**: Now a thin wrapper around CalendarContainerCore that adds swipe paging functionality
- **CalendarBody**: Enhanced to accept `ScrollViewComponent` prop for flexibility

### Build & Export
- Updated rollup.config.js to build both main and legacy entry points
- Added package.json exports for `/legacy` path
- Updated README.md with LegacyCalendar documentation

## Benefits
- No code duplication - single source of truth in CalendarContainerCore
- Legacy build (build/legacy.js) has zero native dependencies
- Smooth migration path for v4 users
- Users can avoid native library installation issues

## Usage

```typescript
// Legacy version (no native deps required)
import { LegacyCalendar } from 'react-native-big-calendar/legacy'
<LegacyCalendar events={events} height={600} />
```
