+++ rollup.config.js
@@ -1,8 +1,7 @@
 import jsx from 'acorn-jsx'
 import typescript2 from 'rollup-plugin-typescript2'
 
-export default {
-  input: './src/index.ts',
+const baseConfig = {
   acornInjectPlugins: [jsx()],
   plugins: [
     typescript2({
@@ -12,17 +11,43 @@ export default {
     }),
   ],
   external: (id) => !id.startsWith('.') && !id.startsWith('/') && id !== 'tslib',
-  output: [
-    {
-      file: 'build/index.js',
-      format: 'cjs',
-      name: 'react-native-big-calendar',
-      sourcemap: true,
-    },
-    {
-      file: 'build/index.es.js',
-      format: 'es',
-      sourcemap: true,
-    },
-  ],
 }
+
+export default [
+  // Main entry point (with native dependencies)
+  {
+    ...baseConfig,
+    input: './src/index.ts',
+    output: [
+      {
+        file: 'build/index.js',
+        format: 'cjs',
+        name: 'react-native-big-calendar',
+        sourcemap: true,
+      },
+      {
+        file: 'build/index.es.js',
+        format: 'es',
+        sourcemap: true,
+      },
+    ],
+  },
+  // Legacy entry point (without native dependencies)
+  {
+    ...baseConfig,
+    input: './src/legacy.ts',
+    output: [
+      {
+        file: 'build/legacy.js',
+        format: 'cjs',
+        name: 'react-native-big-calendar-legacy',
+        sourcemap: true,
+      },
+      {
+        file: 'build/legacy.es.js',
+        format: 'es',
+        sourcemap: true,
+      },
+    ],
+  },
+]
